Moving queries to data access layer
