+++ submit-api/src/submit_api/services/submitted_document_service.py
@@ -1,14 +1,7 @@
 """Service for submitted document management."""
 
-from sqlalchemy import cast, or_, String
-from submit_api.models.account_project import AccountProject
 from submit_api.models.account_project_search_options import DocumentSearchOptions
-from submit_api.models.db import db
-from submit_api.models.item import Item
-from submit_api.models.package import Package
-from submit_api.models.project import Project
-from submit_api.models.submission import Submission, SubmissionType
-from submit_api.models.submitted_document import SubmittedDocument
+from submit_api.models.queries.submitted_document import DocumentQueries
 
 
 class DocumentService:
@@ -17,51 +10,4 @@ class DocumentService:
     @classmethod
     def get_all_documents(cls, search_options: DocumentSearchOptions = None):
         """Get all documents."""
-        session = db.session
-
-        query = session.query(
-            Project.name.label("project_name"),
-            SubmittedDocument.id,
-            SubmittedDocument.name,
-            SubmittedDocument.url,
-            Item.status,
-            # TODO once document version is captured this logic needs to be revisited
-            (cast(Submission.major_version, String) + "." +
-             cast(Submission.minor_version, String)).label("version"),
-            Package.submitted_on
-        )
-
-        # Apply joins step by step
-        query = query.join(AccountProject, AccountProject.project_id == Project.id)
-        query = query.join(
-            Package, Package.account_project_id == AccountProject.id)
-        query = query.join(Item, Item.package_id == Package.id)
-        query = query.join(Submission, Submission.item_id == Item.id)
-        query = query.join(
-            SubmittedDocument, SubmittedDocument.id == Submission.submitted_document_id)
-
-        # Apply filtering conditions
-        query = query.filter(Submission.type == SubmissionType.DOCUMENT)
-
-        # Apply search filters if provided
-        if search_options and any(bool(search_option) for search_option in search_options.__dict__.values()):
-            query = cls.filter_by_search_criteria(query, search_options)
-
-        return query.all()
-
-    @classmethod
-    def filter_by_search_criteria(cls, document_query, search_options: DocumentSearchOptions):
-        """Apply various filters based on search options."""
-
-        if search_options.search_text:
-            document_query = cls._filter_by_search_text(document_query, search_options.search_text)
-
-    @classmethod
-    def _filter_by_search_text(cls, query, search_text):
-        """Filter by search text across package name."""
-        return query.filter(
-            or_(
-                Package.name.ilike(f"%{search_text}%"),
-                Project.name.ilike(f"%{search_text}%")
-            )
-        )
+        return DocumentQueries.get_filtered_documents(search_options)
